requirements.txt
tensorflow-gpu==1.13.1
tensorflow_hub==0.4.0
pillow
piexif